Use virt-preview repo for virtualisation packages
That repository is updated more frequently than other packages in base fedora
and it is done the same way by the same person, so the only difference with this
repository is that the packages are closer to upstream.

Currently libvirt is there in version 4.2.0 (for fedora 28) even though 4.4.0
was released few days ago, so it's not the latest release.  Without this
repository libvirt will be installed in version 4.1.0 on fedora 28 and most
probably not move until there's a new fedora release to adopt.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>