Unify Dockerfiles
The x86_64 and ppc64 Dockerfiles are identical except for two
small details:

  * the path to the QEMU binary, since it encodes the name of the
    architecture;

  * the base image, which in ppc64's case once again contains the
    name of the architecture.

The latter is handled automatically by Docker these days, and the
former can be addressed with a small amount of shell scripting.

By having a single Dockerfile we can drop some code from the
Travis CI configuration and make further changes more convenient,
as well as ensuring architectures are always in sync going
forward.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>