config: Add LIBVIRT_VERSION
As a general rule, we want image creation to be predictable: in
particular, we want the resulting image to contain the exact
version of libvirt we expect.

What we're doing right now is asking for a specific upstream
version (5.0.0) and then tagging the image with the downstream
version available at the time (5.0.0-2.fc30), but a better
approach is to ask for a specific downstream version upfront: if
that version can't be installed the entire process will fail,
prompting the developer to look for another suitable downstream
version and update the configuration file accordingly.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>