config: Add FEDORA_VERSION
Now that we have most information about the environment stored in
the configuration file, we might as well add the Fedora version
too: this way, updating the environment will always require
touching a single file.

Moreover, since the libvirt and QEMU versions include the Fedora
version, we can make the dependency explicit and ensure we don't
accidentally end up installing eg. a QEMU package built for
Fedora 29 inside a container based on Fedora 30.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>