deploy: Change tag format
Right now we're using the libvirt version as the tag, but that's
problematic for a couple of reasons: first of all, other information
such as the QEMU version are not included, which means that even if
we change the contents of the container significantly, as long as
we use the same libvirt version, we end up overwriting the same tag
with every subsequent build; more importantly, it does not include
the architecture name, so the Travis CI job that takes the longest
ends up overwriting the tag created by the one that finished first.

The new scheme is completely disconnected from the libvirt version
and contains enough information to uniquely identify a build: the
shortened git commit hash, along with the architecture name. In
addition to that, the day of the last git update is included as
well, giving images a natural ordering that can easily be taken
advantage of with tools such as sort(1).

The one scenario where the new versioning scheme partially breaks
down is when more than one git change is committed per day, but
that should be rare enough that it doesn't make sense to include
extra information such as a more detailed timestamp to account
for it.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>