deploy: Create and push manifest
Manifests are basically lists of images, and when asked to pull
the Docker client will automatically pick the one that matches the
host architecture.

In addition to creating a manifest for the unique tag that's
associated with the most recent git commit, we also update the
"latest" tag so that it always points to the freshest images.

Right now we hardcode the names of the architectures that are
involved in the process: we will probably want to make things more
generic in the future, but for the time being the current
implementation will do just fine.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>