deploy: Clean up tags
As part of deployment, we push a couple of temporary tags that
we don't want to stick around, so we should remove them at the
end of a successful build.

We use a separate script to remove tags: this provides better
encapsulation, and also stops the Docker Hub token generated in
the process from leaking.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>