deploy: Refactor logic
Up until now we've hardcoded a lot of information regarding what
architectures we want to build on and which of the corresponding
workers is tasked with performing additional work; as a result,
we've made it more difficult to add or (perhaps temporarily)
remove support for certain architectures.

This commit changes the deploy logic so that updating the list
of architectures we want to build on requires touching only two
configuration files and zero code.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>