Remove override of default user/group values
By default, libvirt runs with user and group as 'qemu' with
gid and uid as 107. In case of DPDK deployments, in order to
share the vhostuser sockets, a common group will be used to
run both libvirt and openvswitch, which is 'hugetlbfs'.

In the current impementation, it is required to re-write
the qemu.conf file to change this configuration. Since the
default value of libvirt is 'qemu', removing this value will
not change any behavior. And it will allow kubevirt to append
group configuration, only for OvS-DPDK.

Signed-off-by: Saravanan KR <skramaja@redhat.com>