Remove override of default user/group values
By default, libvirt runs with user and group as 'qemu' with
gid and uid as 107. In case of DPDK deployments, in order to
share the vhostuser sockets, a common group will be used to
run both libvirt and openvswitch, which is 'hugetlbfs'.

Allow the users of libvirt container to override the group
value, like kubevirt can override it based on the deployment
configuration (only applied nodes which requires OvS-DPDK).

Signed-off-by: Saravanan KR <skramaja@redhat.com>