Drop libvirtd.sh
Because this container is used as a base for virt-handler
in kubevirt, which runs as pid 1, this entrypoint is
never called. Even if it was, most if not all of the
config in here is obsolete:

* kubevirt 4a25a1c3 moved much of the config to virt-handler code
* kubevirt 6bd18b3a removed some of the /sys.X usage
* kubevirt 5113f587 removed LIBVIRTD_DEFAULT_NETWORK_DEVICE usage,
  and it's irrelevant nowadays since the libvirt network driver
  isn't installed

This drops the libvirtd.sh script entirely. This also drops ENTRYPOINT
from the docker image entirely, since AFAIK it's not intended to be
usefully run as a standalone service. But maybe I'm wrong about that

Much of the README.md content about running this libvirtd instance
as though it was your host system libvirtd instance is no longer
relevant. This also drops that content

Signed-off-by: Cole Robinson <crobinso@redhat.com>