dockerfile: Create compatibility symlink
Before switching to vanilla Fedora packages, we were using
rebuilt RHEL content: one of the main differences is that with
the former QEMU is installed as

  /usr/bin/qemu-system-x86_64

whereas the latter installs it as

  /usr/libexec/qemu-kvm

This causes problems when migrating VMs from older versions of
KubeVirt, because libvirt will be unable to find the binary in
the old path. Creating a compatibilty symlink works around the
issue.

https://github.com/kubevirt/kubevirt/issues/4629

Signed-off-by: Andrea Bolognani <abologna@redhat.com>