Support rebuilt packages
If the SRPM doesn't come from the same OS that's used for the
container image, then the last part of the version number will not
match.

We've been able to ignore this fact so far because the COPR names
didn't include the full version number, but that has caused issues
in the past already so all future COPRs will include the full
version number in their names.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>